Add c-aci-attestation library (#28)
### Why

This tool will allow us to easily fetch attestations in the clients and
verify those attestations in the server. The tool will eventually be
spun out into it's own repository as it has uses beyond this project

### How

c-aci-attestation is a C library with bindings to support multiple
languages, and a docker image, allowing the user to choose which
packaging makes most sense for their project.

This is a minimal version which just supports C and docker, and only
supports the attestation format expected by CCF

- Update the devcontainer with the required dependencies
- Add CI testing for c-aci-attestation
- Update virtual_attestation_client example container to use
c-aci-attestation (relying on code being local)
- Add c-aci-attestation under tools/ which currently has 3 entry point
executables
  - Get SNP version (SEV, SEV_GUEST or VIRTUAL)
  - Get Attestation in a CCF format
  - Verify Attestation in a CCF format

### TODO

- [ ] Integrate into confidential-server #32 
  - [x] Perhaps via a python binding
- [ ] Add tests with CCF to ensure the CCF format is correct #33